Significantly improved query times for the server list.
Instead of fetching a lit of server IDs it will now fetch the servers in
batches of 50 (by default). In addition, threading has been added at the
server level to speed up the process of querying information of all servers.

Ideally information of the servers should be cached, but since we want up to
date information about player count there is no obvious way to do this. It may
however be an idea to implement some short caching (like, 60s) in case this
would see a high amount of traffic.

Also moved some static configuration to config.json.